%     The multitaper R package
%     Multitaper and spectral analysis package for R
%     Copyright (C) 2011 Karim Rahim 

%     This file is part of the multitaper package for R.

%     The multitaper package is free software: you can redistribute it and
%     or modify
%     it under the terms of the GNU General Public License as published by
%     the Free Software Foundation, either version 2 of the License, or
%     any later version.

%     The multitaper package is distributed in the hope that it will be 
%     useful, but WITHOUT ANY WARRANTY; without even the implied warranty 
%     of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
%     GNU General Public License for more details.

%     You should have received a copy of the GNU General Public License
%     along with multitaper.  If not, see <http://www.gnu.org/licenses/>.

%     If you wish to report bugs please contact the author. 
%     karim.rahim@gmail.com
%     112 Jeffery Hall, Queen's University, Kingston Ontario
%     Canada, K7L 3N6

\name{centralEnglandDailyTemp}
\alias{centralEnglandDailyTemp}
\title{Central England Daily temperature time series}
\description{
  Central England monthly temperature time serie. The data set represents
  daily CET values expressed in Celcius . The data set
  contains temperatures for the years 1772 through 2010.  This dataset was
  retrieved from the Met office at Hadley on Oct 15, 2011.
}
\usage{centralEnglandDailyTemp}
\format{  A data frame indicating the year, month, and temperature from
     observed in Central England. This data-set contains 4224
     observations.}
\keyword{datasets}
