%     The multitaper R package
%     Multitaper and spectral analysis package for R
%     Copyright (C) 2011 Karim Rahim 
%
%     Written by Karim Rahim and Wesley Burr.
%
%     This file is part of the multitaper package for R.
%
%     The multitaper package is free software: you can redistribute it and
%     or modify it under the terms of the GNU General Public License as 
%     published by the Free Software Foundation, either version 2 of the 
%     License, or any later version.
%
%     The multitaper package is distributed in the hope that it will be 
%     useful, but WITHOUT ANY WARRANTY; without even the implied warranty 
%     of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
%     GNU General Public License for more details.
%
%     You should have received a copy of the GNU General Public License
%     along with multitaper.  If not, see <http://www.gnu.org/licenses/>.
%
%     If you wish to report bugs please contact the author. 
%     karim.rahim@gmail.com
%     112 Jeffery Hall, Queen's University, Kingston Ontario
%     Canada, K7L 3N6

\name{CETdaily}
\alias{CETdaily}
\title{Central England Temperature daily time series}
\description{
  Central England Temperature daily time series from the Hadley Center United Kingdom
  Meteorological Office, http://www.metoffice.gov.uk/hadobs/hadcet/.
  The data set represents daily CET values recorded to the tenth of a degree Celsius, 
  and contains temperatures for the years 1772 through September 30, 2011.
  This dataset was retrieved from the Met office at Hadley on Oct 1, 2011.
}
\usage{CETdaily}
\format{ A data frame indicating the year, month, day, and temperature 
     observed in Central England. This data-set contains 87566
     observations.}
\keyword{datasets}
