%     The multitaper R package
%     Multitaper and spectral analysis package for R
%     Copyright (C) 2011 Karim Rahim 
%
%     Written by Karim Rahim and Wesley Burr.
%
%     This file is part of the multitaper package for R.
%
%     The multitaper package is free software: you can redistribute it and
%     or modify it under the terms of the GNU General Public License as 
%     published by the Free Software Foundation, either version 2 of the 
%     License, or any later version.
%
%     The multitaper package is distributed in the hope that it will be 
%     useful, but WITHOUT ANY WARRANTY; without even the implied warranty 
%     of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
%     GNU General Public License for more details.
%
%     You should have received a copy of the GNU General Public License
%     along with multitaper.  If not, see <http://www.gnu.org/licenses/>.
%
%     If you wish to report bugs please contact the author. 
%     karim.rahim@gmail.com

\name{HadCRUTnh}
\alias{HadCRUTnh}
\title{HadCRUT Land Temperature Anomaly (Northern Hemisphere) Series}
\description{
  Hadley Climate Research Unit Temperature anomaly (Northern Hemisphere) 
  time series. Consists of monthly observations, truncated to start at
  March 1958 and extend to September 2009, to match mlco2 dataset.
  This dataset was retrieved from the Hadley CRU on Oct 1, 2011.
}
\usage{HadCRUTnh}
\format{ A data frame indicating the year, month and temperature anomaly 
      for the Northern Hemisphere, between 1958 and 2009.}
\keyword{datasets}
