%     The multitaper R package
%     Multitaper and spectral analysis package for R
%     Copyright (C) 2011 Karim Rahim 
%
%     Written by Karim Rahim and Wesley Burr.
%
%     This file is part of the multitaper package for R.
%
%     The multitaper package is free software: you can redistribute it and
%     or modify it under the terms of the GNU General Public License as 
%     published by the Free Software Foundation, either version 2 of the 
%     License, or any later version.
%
%     The multitaper package is distributed in the hope that it will be 
%     useful, but WITHOUT ANY WARRANTY; without even the implied warranty 
%     of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
%     GNU General Public License for more details.
%
%     You should have received a copy of the GNU General Public License
%     along with multitaper.  If not, see <http://www.gnu.org/licenses/>.
%
%     If you wish to report bugs please contact the author. 
%     karim.rahim@gmail.com

\name{mlco2}
\alias{mlco2}
\title{Mauna Loa Observatory CO2 Monthly Averages}
\description{
  Observations of monthly CO2 atmospheric concentration averages from the 
  Mauna Loa Observatory, Mauna Loa, Hawaii, USA. Obtained from the ESRL
  Global Monitoring Division of the National Oceanic and Atmospheric Administration
  at http://www.esrl.noaa.gov/gmd/dv/data/index.php?parameter_name=Carbon\%2BDioxide. 
  Dataset downloaded Oct 1, 2011.
}
\usage{mlco2}
\format{ A data frame indicating the year, month and atmospheric concentration 
   of CO2 in PPM.}
\keyword{datasets}
