%     The multitaper R package
%     Multitaper and spectral analysis package for R
%     Copyright (C) 2013 Karim Rahim 
%
%     Written by Karim Rahim and Wesley Burr.
%
%     This file is part of the multitaper package for R.
%
%     The multitaper package is free software: you can redistribute it and
%     or modify it under the terms of the GNU General Public License as 
%     published by the Free Software Foundation, either version 2 of the 
%     License, or any later version.
%
%     The multitaper package is distributed in the hope that it will be 
%     useful, but WITHOUT ANY WARRANTY; without even the implied warranty 
%     of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
%     GNU General Public License for more details.
%
%     You should have received a copy of the GNU General Public License
%     along with multitaper.  If not, see <http://www.gnu.org/licenses/>.
%
%     If you wish to report bugs please contact the author. 
%     karim.rahim@gmail.com
%     112 Jeffery Hall, Queen's University, Kingston Ontario
%     Canada, K7L 3N6


\name{percivalAR4}
\alias{percivalAR4}
\title{Auto Regressive Series generated by Don Percival at Applied
  Physics Laboratory (DON PERCIVAL)}
\description{This is a simulated AR(4) time series (page 45 of Percival
  and Walden 1993). The source of this series is: Applied Physics
  Laboratory (Don Percival). The value for delta T is 1, and the sample size is 1024.
}
\usage{percivalAR4}
\format{A time series object containing 1024 simulated values.}   
\references{
Percival, D.B. and Walden, A.T. (1993)
\emph{Spectral analysis for physical applications.}
Cambridge University Press.
}
\examples{
data(percivalAR4)
}
\keyword{datasets}
